requirements.txt
Django==5.0.1

django-extensions==3.2.3
djangorestframework==3.14.0

black==23.12.1
flake8==7.0.0
isort== 5.13.2
